% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fun.arg.default}
\alias{fun.arg.default}
\title{Change Function Defaults.}
\usage{
fun.arg.default(func, ...)
}
\arguments{
\item{func}{the function to have it's defaults changed.}

\item{...}{the new defaults mapping.}
}
\value{
A remapped function.
}
\description{
Allows for the changing of function default arguments.
}
\details{
For every pair of the form \code{name=value} in \code{...}, the function changes the default value of the argument with name \code{name} to the given value \code{value}.
}
