% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fun.res.remap}
\alias{fun.res.remap}
\title{Remaps a functions return result format using the given mapping.}
\usage{
fun.res.remap(func, mapping, drop = FALSE)
}
\arguments{
\item{func}{the function to have it's result format remapped.}

\item{mapping}{the mapping to remap the function with.}

\item{drop}{if return values that weren't in the mapping should be dropped.}
}
\value{
A remapped function.
}
\description{
Like the select function, but does not drop unmatched results.
}
\details{
For every pair in the mapping of for \code{(x='y')}, the function take the result of name x and renames it to y.
}
