% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fun.res.select}
\alias{fun.res.select}
\title{Filters a function to return only values matching the given selection.}
\usage{
fun.res.select(func, selection)
}
\arguments{
\item{func}{the function to wrap.}

\item{selection}{the results to be retained from the function.}
}
\description{
Filters a function to return only values matching the given selection.
}
