% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.boolean}
\alias{is.boolean}
\alias{is.boolean}
\alias{is.true}
\alias{is.boolean}
\alias{is.false}
\title{Boolean Value Test}
\usage{
is.boolean(x)

is.true(x)

is.false(x)
}
\arguments{
\item{x}{the value to test}
}
\value{
\code{is.boolean(x)} \code{TRUE} if the value is a boolean, \code{FALSE} otherwise.

\code{is.true(x)} \code{TRUE} if the value is true, \code{FALSE} otherwise.

\code{is.false(x)} \code{TRUE} if the value is false, \code{FALSE} otherwise.
}
\description{
The following methods were added as to increase consistency with other is.something methods.

Test if a Value is TRUE

Test if a Value is FALSE
}
