% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.numeric.na.null}
\alias{is.numeric.na.null}
\title{Determines if a value is numeric, na, or null.}
\usage{
\method{is.numeric}{na.null}(x)
}
\arguments{
\item{x}{the value to test.}
}
\description{
Uses is.numeric(), is.na(), and is.null() internally.
}
\details{
Have to give export name due to oddity with roxygen export.
}
