% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{named}
\alias{named}
\title{Utility Value Naming Function}
\usage{
named(value, ...)
}
\arguments{
\item{value}{the value(s) to be named.}

\item{...}{the name to(s) give the value(s).}
}
\value{
the named value(s).
}
\description{
Utility function to add names to a given value.
}
\details{
Implemented as:

\code{\link[base]{names}(value) <- \link[base]{list}(...)}
}
\seealso{
\code{\link[base]{unname}}
}
