% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{printf}
\alias{printf}
\title{The missing printf function.}
\usage{
printf(format, ...)
}
\arguments{
\item{format}{the format string to use.}

\item{...}{the arguments for the given format string.}
}
\description{
The missing printf function.
}
