% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdist.R
\name{qdelta.list}
\alias{qdelta.list}
\title{Compute the Delta Between Two Lists.}
\usage{
qdelta.list(x, y)
}
\arguments{
\item{x}{the first list.}

\item{y}{the second list.}
}
\value{
A single numeric delta value.
}
\description{
Computes the delta between lists, \code{x} and \code{y}.
}
\details{
Computes the delta between each value pair in the two lists, then sums as to get the final delta value.
}
\seealso{
Other qdelta functions: \code{\link{qdelta.number}},
  \code{\link{qdelta.numeric}},
  \code{\link{qdelta.string}}, \code{\link{qdelta}}
}
