% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdist.R
\name{qdelta.number}
\alias{qdelta.number}
\alias{qdelta.boolean}
\alias{qdelta.number}
\title{Compute the Delta Between Two Numbers}
\usage{
qdelta.boolean(x, y)

qdelta.number(x, y)
}
\arguments{
\item{x}{the first boolean value.}

\item{y}{the second boolean value.}

\item{x}{the first number.}

\item{y}{the second number.}
}
\value{
A single numeric delta value.

A single numeric delta value.
}
\description{
Computes the delta between two boolean values, \code{x} and \code{y}.

Computes the delta between two numbers, \code{x} and \code{y}.
}
\details{
Returns 0 or 1 depending on if the two boolean values are different. In the case that something went wrong, will return \code{Inf}.

Computes the absolute value of the difference between the two numbers.
}
\seealso{
Other qdelta functions: \code{\link{qdelta.list}},
  \code{\link{qdelta.numeric}},
  \code{\link{qdelta.string}}, \code{\link{qdelta}}

Other qdelta functions: \code{\link{qdelta.list}},
  \code{\link{qdelta.numeric}},
  \code{\link{qdelta.string}}, \code{\link{qdelta}}
}
