% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdist.R
\name{qdelta.numeric}
\alias{qdelta.numeric}
\title{Compute the Delta Between Two Numeric Values}
\usage{
qdelta.numeric(x, y)
}
\arguments{
\item{x}{the first numeric value.}

\item{y}{the second numeric value.}
}
\value{
A single numeric delta value.
}
\description{
Computes the delta between two numeric values, \code{x} and \code{y}.
}
\details{
Computes the euclidean distance between the numeric values, which are possibly numeric lists.
}
\seealso{
Other qdelta functions: \code{\link{qdelta.list}},
  \code{\link{qdelta.number}}, \code{\link{qdelta.string}},
  \code{\link{qdelta}}
}
