% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdist.R
\name{qdelta}
\alias{qdelta}
\title{Compute the Delta Between Two Values}
\usage{
qdelta(x, y)
}
\arguments{
\item{x}{the first value.}

\item{y}{the second value.}
}
\value{
A single numeric delta value.
}
\description{
Computes the delta between two alike-type values, \code{x} and \code{y}.
}
\details{
First determines the type of the two values, then calls the respective qdelta function to compute the actual delta value.
}
\seealso{
Other qdelta functions: \code{\link{qdelta.list}},
  \code{\link{qdelta.number}},
  \code{\link{qdelta.numeric}}, \code{\link{qdelta.string}}
}
