% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdist.R
\name{qdelta.string}
\alias{qdelta.string}
\title{Compute the Delta Between Two Strings}
\usage{
qdelta.string(x, y)
}
\arguments{
\item{x}{the first string.}

\item{y}{the second string.}
}
\value{
A single numeric delta value.
}
\description{
Computes the delta between two strings, \code{x} and \code{y}.
}
\details{
Uses \code{\link[utils]{adist}} to compute the string delta value.
}
\seealso{
Other qdelta functions: \code{\link{qdelta.list}},
  \code{\link{qdelta.number}},
  \code{\link{qdelta.numeric}}, \code{\link{qdelta}}
}
