% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdiff.R
\name{qdiff}
\alias{qdiff}
\alias{qdiff.value}
\alias{qdiff}
\alias{qdiff.cell}
\alias{qdiff}
\alias{qdiff}
\alias{vqdiff}
\title{Compute a Difference}
\usage{
qdiff.value(actual, expected, tolerance)

qdiff.cell(row, col, actual, expected, tolerance)

qdiff(actual, expected, tolerances = NULL, drop = TRUE)

vqdiff(actual, expected, tolerances = NULL, drop = TRUE)
}
\arguments{
\item{actual}{the actual value.}

\item{expected}{the expected value.}

\item{tolerance}{the error tolerance.}

\item{row}{the row that the cell is a member of.}

\item{col}{the column that the cell is a member of within the given row.}

\item{tolerances}{a list of tolerance values, on a per-column basis.}

\item{drop}{whether or not to keep information on cells that are determined to be the same.}
}
\value{
a dataframe consisting of all relevant diff data.
}
\description{
Computes the difference between two values.
}
