% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdist.R
\name{qdist}
\alias{qdist}
\alias{qdist.number}
\alias{qdist}
\alias{qdist.numeric}
\alias{qdist}
\alias{qdist.string}
\alias{qdist}
\alias{qdist.list}
\alias{qdist}
\title{Quick Distance}
\usage{
qdist.number(x, y)

qdist.numeric(x, y)

qdist.string(x, y)

qdist.list(x, y)

qdist(x, y)
}
\description{
Computes the distance between two numbers.
}
