% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgen.R
\name{qgen.element}
\alias{qgen.element}
\alias{qgen.element}
\alias{qgen.element.default}
\alias{qgen.element}
\alias{qgen.element.diagonal}
\title{Creates a value to be used in a matric.}
\usage{
qgen.element.default(row, col, nrow, ncol)

qgen.element.diagonal(row, col, nrow, ncol, value = 1, orelse = 0)
}
\arguments{
\item{value}{the value to be placed on the diagonal.}

\item{orelse}{the value to be placed everwhere else.}
}
\value{
a value representing the row * ncol + col.
}
\description{
Creates a value for (row, col) such that value = row * ncol + col.
}
\details{
The default implementation returns the row * number of columns plus the column number.

The diagonal implementation returns a matrix with the given value on the diagonal.
}
