% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgen.R
\name{qgen.matrix.element.default}
\alias{qgen.matrix.element.default}
\alias{qgen.matrix.element.default}
\alias{qgen.matrix.element.diagonal}
\title{Creates a value to be used in a matric.}
\usage{
qgen.matrix.element.default(row, col, nrow, ncol)

qgen.matrix.element.diagonal(row, col, nrow, ncol, value = 1, orelse = 0)
}
\value{
a value representing the row * ncol + col.

a value representing the row * ncol + col.
}
\description{
Creates a value for (row, col) such that value = row * ncol + col.

Creates a value for (row, col) such that value = row * ncol + col.
}
