% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgen.R
\name{qgen}
\alias{qgen}
\alias{qgen.matrix}
\alias{qgen}
\alias{qgen.data.frame}
\title{Create a nrow by ncol matrix.}
\usage{
qgen.matrix(nrow, ncol, func = qgen.matrix.element.default, ...)

qgen.data.frame(nrow, ncol, func = qgen.matrix.element.default, ...)
}
\arguments{
\item{nrow}{the number of rows that will be in the resulting matrix.}

\item{ncol}{the number of columns that will be int the resulting matrix.}

\item{func}{the function used to generate values.}
}
\value{
a matrix generated with the given generator function.
}
\description{
Creates a matrix using the given generator function.

Creates a matrix using the given generator function.
}
