% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quikr.R
\docType{package}
\name{quikr}
\alias{quikr}
\alias{quikr-package}
\alias{quikr-package}
\title{QuikR: Assorted R Utility Functions}
\description{
QuikR is a package containing assorted utility functions for use in the R language.
}
\details{
Nothing to see here atm.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rlonryan/QuikR}
}

}
