% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{remap}
\alias{remap}
\title{Remaps a functions return result format using the given mapping.}
\usage{
remap(func, mapping, drop = FALSE)
}
\arguments{
\item{func}{the function to have it's result format remapped.}

\item{mapping}{the mapping to remap the function with.}
}
\description{
Like the select function, but does not drop unmatched results.
}
