% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablediff.R
\name{table.diff}
\alias{table.diff}
\alias{table.vdiff}
\title{Computes the difference of two values, with the given precision.}
\usage{
table.diff(a, e, digits = 4, incl.a = FALSE, incl.e = FALSE,
  incl.d = TRUE, drop = FALSE)

table.vdiff(a, e, ...)
}
\arguments{
\item{a}{the actual value.}

\item{e}{the expected value.}

\item{digits}{the number of decimal places to round values to.}

\item{incl.a}{if the acutal   value for each row should be included in the result.}

\item{incl.e}{if the expected value for each row should be included in the result.}

\item{incl.d}{if the delta    value for each row should be included in the result.}

\item{drop}{if non-differing rows should be dropped from the result.}
}
\description{
Computes the difference of two values, with the given precision.

Diffs two tables and opens the result in a viewer.
}
