% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdiff.R
\name{vqdiff}
\alias{vqdiff}
\alias{vqdiff.value}
\alias{vqdiff}
\alias{vqdiff.value}
\alias{vqdiff}
\alias{vqdiff.cell}
\alias{vqdiff}
\alias{vqdiff.table}
\title{Views the Diff of Two Values}
\usage{
vqdiff.value(actual, expected, tolerance)

vqdiff.cell(row, col, actual, expected, tolerance)

vqdiff.table(actual, expected, tolerances = NULL, drop = TRUE)
}
\arguments{
\item{actual}{the actual value.}

\item{expected}{the expected value.}

\item{tolerance}{the error tolerance.}

\item{row}{the row that the cell is a member of.}

\item{col}{the column that the cell is a member of within the given row.}

\item{tolerances}{a list of tolerance values, on a per-column basis.}

\item{drop}{whether or not to keep information on cells that are determined to be the same.}
}
\value{
A dataframe consisting of the diff data.
}
\description{
Computes the difference between two values, \code{actual} and \code{expected}, and then views it.
}
\details{
Determines the difference using \code{\link[quikr]{qdelta}}, and compares against the given tolerance level.
}
\seealso{
\code{\link[quikr]{qdiff}}
}
