% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fun.res.select}
\alias{fun.res.select}
\title{Select Function Return Values}
\usage{
fun.res.select(func, selection)
}
\arguments{
\item{func}{the function to wrap.}

\item{selection}{the results to be retained from the function.}
}
\value{
A wrapped function.
}
\description{
Filters a function to return only values matching the given selection.
}
\details{
For a given function, the function returns a wrapper that only retains return values specified in the selected list.
}
