% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{named}
\alias{named}
\alias{unnamed}
\title{Utility function to create a named value.}
\usage{
named(value, ...)

unnamed(value)
}
\arguments{
\item{value}{the value(s) to be named.}

\item{...}{the name to(s) give the value(s).}
}
\value{
the named value(s).
}
\description{
Utility function to create a named value.
}
