% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{printf}
\alias{printf}
\title{Formatted Print Function}
\usage{
printf(format, ...)
}
\arguments{
\item{format}{the format string to use.}

\item{...}{the arguments for the given format string.}
}
\description{
The missing printf function.
}
\details{
Uses \code{\link[base]{sprintf}} internally.
}
\seealso{
\code{\link[base]{sprintf}}
}
