% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgen.R
\name{qgen}
\alias{qgen}
\alias{qgen.matrix}
\alias{qgen}
\alias{qgen.matrix}
\alias{qgen}
\alias{qgen.data.frame}
\title{Generates a Table}
\usage{
qgen.matrix(nrow, ncol, func = qgen.element.default, ...)

qgen.data.frame(nrow, ncol, func = qgen.element.default, ...)
}
\arguments{
\item{nrow}{the number of rows that will be in the resulting matrix.}

\item{ncol}{the number of columns that will be int the resulting matrix.}

\item{func}{the function used to generate values.}
}
\value{
a table generated with the given generator function.
}
\description{
Creates a table with dimensions \code{nrow} x \code{ncol} using the given generator function.
}
\details{
Generates a table (either a matrix or a dataframe), using the given generator function, passing any \code{...} arguments to the generator function.
}
