% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{yesno}
\alias{yesno}
\title{Convert a Boolean to a String}
\usage{
yesno(bool, yes = "yes", no = "no")
}
\arguments{
\item{bool}{the boolean value to test.}

\item{yes}{the value to return on a true  input.}

\item{no}{the value to return on a false input.}
}
\value{
'yes' or 'no'.
}
\description{
Returns 'yes' or 'no' based off of the provided boolean input.
}
\details{
If \code{yes} is specified, the value of \code{yes} is used instead of the default value \code{'yes'}.

If \code{no} is specified, the value of \code{no} is used instead of the default value \code{'no'}.
}
