% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{yesno}
\alias{yesno}
\alias{yn}
\title{Returns 'yes' or 'no' based off of the provided boolean input.}
\usage{
yn(bool, yes = "yes", no = "no")
}
\arguments{
\item{yes}{the value to return on a true  input.}

\item{no}{the value to return on a false input.}
}
\description{
Returns 'yes' or 'no' based off of the provided boolean input.
}
